requirements.txt
groq>=0.13.0
langchain>=0.3.7
langchain-groq>=0.2.1
langgraph>=0.2.45
tavily-python>=0.5.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.0
reportlab>=4.0.0
pydantic>=2.0.0
sse-starlette>=2.0.0
aiofiles>=24.0.0
